refactor: enable pruning from relevant stores
