--- src/snapshot/mod.rs
@@ -1,3 +1,0 @@
-pub mod snapshot_creator;
-pub mod timeframe_change_handler;
-pub mod utils;
