+++ src/timeframe_management/utils.rs
@@ -1,10 +1,6 @@
-use std::{cell::RefCell, rc::Rc};
-
 use substreams::store::{DeltaInt64, Deltas};
 
-use crate::{key_management::store_key_manager::StoreKey, types::timeframe::Timeframe};
-
-use super::snapshot_creator::SnapshotCreator;
+use crate::key_management::store_key_manager::StoreKey;
 
 pub fn separate_timeframe_deltas(
     deltas: &Deltas<DeltaInt64>,
@@ -29,24 +25,6 @@ pub fn separate_timeframe_deltas(
             deltas: hourly_deltas,
         },
     )
-}
-
-pub fn prepare_snapshot_closure(
-    snapshot_creator: Rc<RefCell<SnapshotCreator>>,
-    snapshot_type: Timeframe,
-) -> impl FnMut(i64) + '_ {
-    move |time_frame_id| {
-        let mut creator = snapshot_creator.borrow_mut();
-        match snapshot_type {
-            Timeframe::Daily => {
-                creator.create_protocol_financials_daily_snapshot(&time_frame_id);
-                creator.create_liquidity_pool_snapshots(&snapshot_type, &time_frame_id);
-            }
-            Timeframe::Hourly => {
-                creator.create_liquidity_pool_snapshots(&snapshot_type, &time_frame_id);
-            }
-        }
-    }
 }
 
 pub fn calculate_day_hour_id(timestamp_seconds: i64) -> (i64, i64) {
